+++ js/listener.js
@@ -65,12 +65,20 @@
     };
 
     OCA.Onlyoffice.onDocumentReady = function (documentType) {
-        if (documentType === "text") {
+        if (documentType === "word") {
             OCA.Onlyoffice.bindVersionClick();
         } else {
             OCA.Onlyoffice.unbindVersionClick();
         }
     };
+
+    OCA.Onlyoffice.changeFavicon = function (favicon) {
+        $('link[rel="icon"]').attr("href", favicon);
+    };
+
+    OCA.Onlyoffice.onShowMessage = function (messageObj) {
+        OC.Notification.show(messageObj.message, messageObj.props);
+    }
 
     window.addEventListener("message", function (event) {
         if ($("#onlyofficeFrame")[0].contentWindow !== event.source
@@ -84,6 +92,11 @@
             case "editorRequestSharingSettings":
                 if (OCA.Onlyoffice.OpenShareDialog) {
                     OCA.Onlyoffice.OpenShareDialog();
+                }
+                break;
+            case "onRefreshVersionsDialog":
+                if (OCA.Onlyoffice.RefreshVersionsDialog) {
+                    OCA.Onlyoffice.RefreshVersionsDialog();
                 }
                 break;
             case "editorRequestSaveAs":
@@ -101,12 +114,25 @@
             case "onDocumentReady":
                 OCA.Onlyoffice.onDocumentReady(event.data.param);
                 break;
+            case "changeFavicon":
+                OCA.Onlyoffice.changeFavicon(event.data.param);
+                break;
+            case "onShowMessage":
+                OCA.Onlyoffice.onShowMessage(event.data.param);
         }
     }, false);
 
     window.addEventListener("popstate", function (event) {
-        if ($("#onlyofficeFrame").length) {
+        if ($("#onlyofficeFrame").length
+            && location.href.indexOf(OCA.Onlyoffice.AppName) == -1) {
             OCA.Onlyoffice.onRequestClose();
+        }
+    });
+
+    window.addEventListener("DOMNodeRemoved", function(event) {
+        if ($(event.target).length && $("#onlyofficeFrame").length
+            && ($(event.target)[0].id === "viewer" || $(event.target)[0].id === $("#onlyofficeFrame")[0].id)) {
+            OCA.Onlyoffice.changeFavicon($("#onlyofficeFrame")[0].contentWindow.OCA.Onlyoffice.faviconBase);
         }
     });