+++ lib/templatemanager.php
@@ -20,6 +20,7 @@
 namespace OCA\Onlyoffice;
 
 use OCP\Files\Folder;
+use OCP\Files\NotFoundException;
 
 /**
  * Template manager
@@ -48,10 +49,15 @@ class TemplateManager {
      * @return Folder
      */
     public static function GetGlobalTemplateDir() {
-        $rootFolder = \OC::$server->getRootFolder();
+        $dirPath = self::$appName . "/" . self::$templateFolderName;
 
-        $appDir = $rootFolder->nodeExists(self::$appName) ? $rootFolder->get(self::$appName) : $rootFolder->newFolder(self::$appName);
-        $templateDir = $appDir->nodeExists(self::$templateFolderName) ? $appDir->get(self::$templateFolderName) : $appDir->newFolder(self::$templateFolderName);
+        $rootFolder = \OC::$server->getRootFolder();
+        $templateDir = null;
+        try {
+            $templateDir = $rootFolder->get($dirPath);
+        } catch (NotFoundException $e) {
+            $templateDir = $rootFolder->newFolder($dirPath);
+        }
 
         return $templateDir;
     }
@@ -66,10 +72,10 @@ class TemplateManager {
     public static function GetGlobalTemplates($mimetype = null) {
         $templateDir = self::GetGlobalTemplateDir();
 
-        if (!empty($mimetype)) {
+        $templatesList = $templateDir->getDirectoryListing();
+        if (!empty($mimetype)
+            && is_array($templatesList) && count($templatesList) > 0) {
             $templatesList = $templateDir->searchByMime($mimetype);
-        } else {
-            $templatesList = $templateDir->getDirectoryListing();
         }
 
         return $templatesList;